api/requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
uvicorn[standard]==0.24.0
pytesseract==0.3.10
pdf2image==1.16.3
openai==1.3.8
fpdf2==2.7.6
PyPDF2==3.0.1
langchain==0.1.0
langchain-openai==0.0.5
langchain-pinecone==0.0.3
pinecone-client==3.0.0
opencv-python-headless==4.8.1.78
numpy==1.24.4
python-dotenv==1.0.0
pillow==10.1.0
pydantic==2.5.0